int conversion

+++ molvae/preprocess.py
@@ -51,7 +51,7 @@ if __name__ == "__main__":
 
     all_data = pool.map(tensorize, data)
 
-    le = (len(all_data) + num_splits - 1) / num_splits
+    le = int((len(all_data) + num_splits - 1) / num_splits)
 
     for split_id in range(num_splits):
         st = split_id * le
